requirements.txt
Flask
SQLAlchemy
python-dotenv